[SALEOR-3384] Setup Apollo config
