+++ src/core/createSaleorClient.ts
@@ -9,8 +9,13 @@ import { Config } from "./types";
  * @param config - Config for creating apollo client.
  */
 export const createSaleorClient = (
-  config: Config
+  config: Config,
+  apolloConfig: ApolloClient<any>
 ): ApolloClient<NormalizedCacheObject> => {
+  if (apolloConfig) {
+    return new ApolloClient(apolloConfig);
+  }
+
   return new ApolloClient({
     uri: config.apiUrl,
     cache: new InMemoryCache(),
