+++ src/core/index.ts
@@ -1,14 +1,2 @@
-import { ApolloClient, NormalizedCacheObject } from "@apollo/client";
-import { auth, AuthSDK } from "./auth";
-
-export interface Core {
-  auth: AuthSDK;
-}
-
-export const SaleorSDK = (
-  client: ApolloClient<NormalizedCacheObject>
-): Core => {
-  const authSDK = auth(client);
-
-  return { auth: authSDK };
-};
+export * from "./SaleorSDK";
+export * from "./createSaleorClient";
