+++ src/react/components/SaleorProvider.tsx
@@ -1,43 +1,37 @@
-import React, { useState, useEffect } from "react";
-import { Config } from "../../core/types";
-import {
-  ApolloClient,
-  ApolloProvider,
-  NormalizedCacheObject,
-} from "@apollo/client";
-import { createSaleorClient } from "../../core/createSaleorClient";
-import { SaleorContext, SaleorContextType } from "../context";
-import { SaleorSDK } from "../../core";
+import React from "react";
+import { ApolloClient } from "@apollo/client";
+import { ApolloProvider } from "@apollo/client/react";
+import { createSaleorClient, SaleorSDK } from "../../core";
 
-export interface SaleorProviderProps {
-  children: React.ReactElement | React.ReactNode;
-  config: Config;
-}
+const SaleorContext = React.createContext<{ api: any; config: any } | null>(
+  null
+);
 
-export const SaleorProvider: React.FC<SaleorProviderProps> = ({
+export const SaleorProvider: React.FC<{ config: any; apolloConfig: any }> = ({
   config,
+  apolloConfig,
   children,
 }) => {
-  const [client, setClient] = useState<ApolloClient<
-    NormalizedCacheObject
-  > | null>(null);
-  const [context, setContext] = useState<SaleorContextType | null>(null);
+  const [client, setClient] = React.useState<ApolloClient<any> | null>(null);
+  const [context, setContext] = React.useState<{
+    api: any;
+    config: any;
+  } | null>(null);
 
-  useEffect(() => {
-    const client = createSaleorClient(config);
+  React.useEffect(() => {
+    const client = createSaleorClient(config, apolloConfig);
     setClient(client);
     const api = SaleorSDK(client);
     setContext({ api, config });
-  }, [config]);
+  }, []);
 
   if (client && context) {
     return (
-      <ApolloProvider client={client}>
-        <SaleorContext.Provider value={context}>
-          {children}
-        </SaleorContext.Provider>
-      </ApolloProvider>
+      <SaleorContext.Provider value={null}>
+        <ApolloProvider client={client}>{children}</ApolloProvider>
+      </SaleorContext.Provider>
     );
   }
+
   return null;
 };
